App is now location-aware, but mapping features are prone to bugs.
For now, in order to view the current location (1 Infinite Loop, Cupertino, CA, in the iOS Simulator), you have to back out of the caravan and re-enter it. This is solely a demonstration of the application's location-awareness...I am currently working on a method that will enable live location updating with annotation support for user pins from the database.